Moved homepage components to `@/app/_components`
