+++ src/app/_components/land.jsx
@@ -1,0 +1,33 @@
+export default function Land() {
+  return (
+    <>
+      <section id="land">
+        <h1 className="text-center text-4xl lg:text-5xl pb-8 main text-highlight-dark my-auto">
+          Land Acknowlegement
+        </h1>
+        <div className="bg-highlight-dark py-2 flex justify-center items-center">
+          <div className="text-container w-[90%] lg:w-[75%] h-full">
+            <p className="text-background py-2 text-left main ">
+              Toronto Metropolitan University is in the ‘Dish With One Spoon
+              Territory’. The Dish With One Spoon is a treaty between the
+              Anishinaabe, Mississaugas and Haudenosaunee that bound them to share
+              the territory and protect the land. Subsequent Indigenous Nations
+              and peoples, Europeans and all newcomers have been invited into this
+              treaty in the spirit of peace, friendship and respect.
+            </p>
+            <p className="text-background py-2 text-lef main ">
+              The &quot;Dish&quot;, or sometimes it is called the &quot;Bowl&quot;, represents what is
+              now southern Ontario, from the Great Lakes to Quebec and from Lake
+              Simcoe into the United States. We all eat out of the Dish, all of
+              us that share this territory, with only one spoon. That means we
+              have to share the responsibility of ensuring the dish is never
+              empty, which includes taking care of the land and the creatures we
+              share it with. Importantly, there are no knives at the table,
+              representing that we must keep the peace.
+            </p>
+          </div>
+        </div>
+      </section>
+    </>
+  );
+}
