+++ src/app/page.js
@@ -1,8 +1,8 @@
-import Parallax from "./_components/homepage/parallax"
-import Pillars from "./_components/homepage/pillars"
-import Land from "./_components/homepage/land"
-import USSTM from "./_components/homepage/usstm"
-import FAQ from "./_components/homepage/faq"
+import Parallax from "./_components/parallax"
+import Pillars from "./_components/pillars"
+import Land from "./_components/land"
+import USSTM from "./_components/usstm"
+import FAQ from "./_components/faq"
 
 export default function Home() {
   return (
